requirements.txt
openai
langchain
aiocache
sqlalchemy
langchain_community
flask[async]
httpx
requests
sentence_transformers
colorlog
python-dotenv